Create extended_euclidean.cpp
The extended Euclidean algorithm is particularly useful when a and b are co-prime since x is the modular multiplicative inverse of a modulo b, and y is the modular multiplicative inverse of b modulo a. Both extended Euclidean algorithms are widely used in cryptography. The computation of the modular multiplicative inverse is an essential step in the RSA public-key encryption algorithm.